backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
SQLAlchemy==1.4.53
python-dotenv==1.0.0
joblib==1.3.2
torch==2.2.0
scikit-learn==1.7.1
requests==2.31.0
werkzeug==2.2.3
